feat: audit item opens
If the user open item this either means:

1. They started reading it, so the item is interesting.

2. It is not clear what the item is about from the title, they opened
it to refresh their memory. This also implies that they forgot about
the item, so that was a good (or too late) time to remind them about
it.

The number of times the item is opened should probably contribute to
the scheduling algorithm, so we collect it as well (if user has opted
in).