fix: try to workaround sign in bug in safari
See https://github.com/firebase/firebase-js-sdk/issues/4256 for the
description of issue and workarounds.

We already use the correct authDomain, so another thing to try is to
use `signInWithRedirect` instead of `signInWithPopup`. If that doesn’t
help, the next option is using google sign in sdk directly and feeding
token to firebase afterwards.